+++ atoms/headerMenuOpenAtom.ts
@@ -1,6 +1,0 @@
-import { atom } from "@/lib/recoil";
-
-export const headerMenuOpenAtom = atom<boolean>({
-  key: "headerMenuOpenAtom",
-  default: false,
-});